Extract out functions for turning package names into WebPack entry
